Move to node_gyp for compiling ozw on windows
I moved the compilation to node_gyp for open zwave.  This lets you
compile the stock openzwave project with whatever version of visual
studio you have for whatever architecture you have.

I also moved the expect location of the open zwave files to a central
location to be more in line with the non-windows environment
(~/AppData/Local/OpenZwave) and this location is stored in an
environment variable.