Properly inherits from EventEmitter
Instead of creating an instance of an EventEmitter, and creating wrapper functions, we copy directly from the prototype.

This has several benefits over how it was before.
For instance, chaining of listener attachments is now allowd.
zwave
  .on('node added', eventHandle)
  .on('value changed', someOtherEventHandle);
This is because the prototype methods on EventEmitter returns 'this'.
But the wrapper functions did not.