Changes to make the windows build compatible with Node 12.
removed wrench with fs-extra because it is long deprecated.
replaced unzip with unzipper because unzip was not compat.

rewrote some of the file system directory lookups to be compatible with the original intent but using fs-extra.. or a custom recursive folder list.
added the package.json file to the temp folder for compat.